Add `Code` for the audit log entry created when the device is reset.
This isn't a valid command code which is probably why it was never
included here. Yubico does however overload the command code field of
the audit log entry used to indicate that the yubihsm has been reset.
Without this change if the reset log entry is in the log and the
`get_log_entries` command is executed it will omit all log entries
w/o reporting an error. Behind the scenes the deserialization logic is
however failing to parse the exceptional event log entry and silently
failing.